[codex-apps] Isolate file parameter schema handling
