+++ codex-rs/codex-mcp/src/codex_apps.rs
@@ -5,6 +5,11 @@
 
 use codex_utils_plugins::mcp_connector::sanitize_name;
 
+mod file_params;
+
+pub use file_params::declared_openai_file_input_param_names;
+pub(crate) use file_params::prepare_openai_file_params_for_model;
+
 pub(crate) fn normalize_codex_apps_tool_title(connector_name: Option<&str>, value: &str) -> String {
     let Some(connector_name) = connector_name
         .map(str::trim)
