+++ codex-rs/codex-mcp/src/lib.rs
@@ -70,10 +70,10 @@ pub use mcp::oauth_login_support_with_http_client;
 pub use mcp::resolve_oauth_scopes;
 pub use mcp::should_retry_without_scopes;
 
+pub use codex_apps::declared_openai_file_input_param_names;
 pub use mcp::McpPermissionPromptAutoApproveContext;
 pub use mcp::mcp_permission_prompt_is_auto_approved;
 pub use mcp::qualified_mcp_tool_name_prefix;
-pub use tools::declared_openai_file_input_param_names;
 
 pub(crate) mod auth_elicitation;
 mod catalog;
