+++ codex-rs/codex-mcp/src/rmcp_client.rs
@@ -21,6 +21,7 @@ use std::time::Instant;
 use crate::codex_apps::normalize_codex_apps_callable_name;
 use crate::codex_apps::normalize_codex_apps_callable_namespace;
 use crate::codex_apps::normalize_codex_apps_tool_title;
+use crate::codex_apps::prepare_openai_file_params_for_model;
 use crate::codex_apps_cache::CodexAppsToolsCacheContext;
 use crate::codex_apps_cache::CodexAppsToolsFetchSource;
 use crate::codex_apps_cache::load_startup_cached_codex_apps_server_info;
@@ -34,8 +35,6 @@ use crate::server::McpServerLaunch;
 use crate::tools::ToolFilter;
 use crate::tools::ToolInfo;
 use crate::tools::filter_tools;
-use crate::tools::supported_openai_file_input_optional_fields;
-use crate::tools::tool_with_model_visible_input_schema;
 use anyhow::Result;
 use anyhow::anyhow;
 use async_channel::Sender;
@@ -621,9 +620,7 @@ fn prepare_codex_apps_tools_for_model(
     tool_plugin_provenance: &ToolPluginProvenance,
 ) -> Vec<ToolInfo> {
     for tool in &mut tools {
-        tool.openai_file_input_optional_fields =
-            supported_openai_file_input_optional_fields(&tool.tool);
-        tool.tool = tool_with_model_visible_input_schema(&tool.tool);
+        prepare_openai_file_params_for_model(tool);
         let plugin_names = match tool.connector_id.as_deref() {
             Some(connector_id) => {
                 tool_plugin_provenance.plugin_display_names_for_connector_id(connector_id)
